frontend/requirements.txt
Pillow==9.5.0
python-multipart==0.0.6
imagehash==4.3.1
numpy==1.23.1
streamlit==1.39.0
